+++ dataland-frontend/src/components/resources/dataTable/MultiLayerDataTableCells.ts
@@ -1,0 +1,31 @@
+import { type DocumentReference } from "@clients/backend";
+import { type DynamicDialogOptions } from "primevue/dynamicdialogoptions";
+import { type MLDTDisplayValue } from "@/components/resources/dataTable/MultiLayerDataTableConfiguration";
+
+export enum MLDTDisplayComponents {
+  StringDisplayComponent = "StringDisplayComponent",
+  DocumentLinkDisplayComponent = "DocumentLinkDisplayComponent",
+  ModalLinkDisplayComponent = "ModalLinkDisplayComponent",
+}
+
+export type MLDTDisplayComponentTypes = {
+  [MLDTDisplayComponents.StringDisplayComponent]: string;
+  [MLDTDisplayComponents.DocumentLinkDisplayComponent]: { label: string; reference: DocumentReference };
+  [MLDTDisplayComponents.ModalLinkDisplayComponent]: {
+    label: string;
+    // Ignored as "any" type comes from DynamicDialog
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    modalComponent: any;
+    modalOptions?: DynamicDialogOptions;
+  };
+};
+
+export type AvailableDisplayValues =
+  | MLDTDisplayValue<MLDTDisplayComponents.StringDisplayComponent>
+  | MLDTDisplayValue<MLDTDisplayComponents.DocumentLinkDisplayComponent>
+  | MLDTDisplayValue<MLDTDisplayComponents.ModalLinkDisplayComponent>;
+
+export interface MLDTDisplayValue<DisplayComponent extends MLDTDisplayComponents> {
+  displayComponent: DisplayComponent;
+  displayValue: MLDTDisplayComponentTypes[DisplayComponent];
+}
